+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/StoneCypher/circular_buffer_js/compare/v1.7.1...HEAD)
 
+- Add tests for set .capacity, fixes #49 [`#49`](https://github.com/StoneCypher/circular_buffer_js/issues/49)
 - draft implementations [`0dd5d0f`](https://github.com/StoneCypher/circular_buffer_js/commit/0dd5d0f1c73b7752f22367f62e771978625e79c3)
 - unit tests [`3d2637f`](https://github.com/StoneCypher/circular_buffer_js/commit/3d2637f7548883daccb26328c698d3bace86306e)
 - SetLength test was wrong: it asserted that length matched what was set, which isn't true when setting a higher size [`60c7819`](https://github.com/StoneCypher/circular_buffer_js/commit/60c781994047822fb263c05764f7d7aa19791bbb)